v11 CategoryParent Ajax

+++ app/Http/Controllers/Admin/CategoryparentController.php
@@ -11,38 +11,48 @@ class CategoryparentController extends Controller
 {
     public function index()
     {
+        return view('auth.admin.Categoryparent.principal');
+    }
+
+    public function lista()
+    {
         $c = CategoryParent::all();
-        return view('auth.admin.Categoryparent.principal')->with('c',$c);
+        return view('auth.admin.Categoryparent.lista')->with('c',$c);
     }
 
 
     public function store(Request $request)
     {
-        $request->validate([
-            'name'=>'required|max:255'
 
-        ]);
+        if($request->ajax()){
+            $request->validate([
+                'name'=>'required|max:255'
+    
+            ]);    
 
-        $newc= new CategoryParent();
-        $newc->name=$request->name;
-        $newc->save();
+            $newc= new CategoryParent();
+            $newc->name=$request->name;
+            $newc->save();
 
-        return redirect()->route('admin.categoriaPadre.principal')->with('crear','ok');
-
+            //return redirect()->route('admin.categoriaPadre.principal')->with('crear','ok');
+            return response($newc);
+        }
     }
 
     public function update(Request $request, CategoryParent $item)
     {
-        
-        $request->validate([
-            'name'=>'required|max:255'
-            
-        ]);
+        if($request->ajax()){
+            $request->validate([
+                'name'=>'required|max:255'
+                
+            ]);
 
-        $item->name=$request->name;
-        $item->save();
+            $item->name=$request->name;
+            $item->save();
 
-        return redirect()->route('admin.categoriaPadre.principal')->with('editar','ok');
+            //return response()->route('admin.categoriaPadre.principal')->with('editar','ok');
+            return response($item);
+        }
 
         
     }
@@ -50,12 +60,13 @@ class CategoryparentController extends Controller
     public function delete(Request $request, $item)
     {
         //dd($movieid);
-        
-        $item=CategoryParent::find($item);
-        $item->delete();
-        
+        if($request->ajax()){
+            $item=CategoryParent::find($item);
+            $item->delete();
+            
+        }
         
         //return redirect()->back()->with('eliminar','oki');
-        return redirect()->route('admin.categoriaPadre.principal')->with('eliminar','oki');
+        //return redirect()->route('admin.categoriaPadre.principal')->with('eliminar','oki');
     }
 }
